task#284: move C7-Models to C7-Plugin
