+++ kadai-adapter-camunda-system-connector/src/main/java/io/kadai/adapter/monitoring/models/OutboxEventCountRepresentationModel.java
@@ -1,4 +1,4 @@
-package io.kadai.adapter.models;
+package io.kadai.adapter.monitoring.models;
 
 public class OutboxEventCountRepresentationModel {
